==> Training-2018/Training-2018/SparkSummitTraining/Spark-ILT/Python/Solutions/Tuning/02 Utility Functions.py <==
# Databricks notebook source
# MAGIC %md-sandbox
# MAGIC <div style="text-align: center; line-height: 0; padding-top: 9px;">
# MAGIC   <img src="https://databricks.com/wp-content/uploads/2018/03/db-academy-rgb-1200px.png" alt="Databricks Learning" style="width: 600px; height: 163px">
# MAGIC </div>

# COMMAND ----------

# MAGIC %md
# MAGIC #Utility Functions
# MAGIC 
# MAGIC **Technical Accomplishments:**
# MAGIC * Discover techniques for counting files
# MAGIC * Develop tools for future diagnostics efforts
# MAGIC * Introduce the lab format

# COMMAND ----------

# MAGIC %md
# MAGIC ##![Spark Logo Tiny](https://files.training.databricks.com/images/105/logo_spark_tiny.png) Getting Started
# MAGIC 
# MAGIC Run the following cell to configure our "classroom."

# COMMAND ----------

# MAGIC %run "../Includes/Classroom Setup"

# COMMAND ----------

# MAGIC %run "../Includes/Initialize Labs"

# COMMAND ----------

# MAGIC %md-sandbox
# MAGIC <div style="display:table; height:600px; width:100%">
# MAGIC   <h2><img src="https://files.training.databricks.com/images/105/logo_spark_tiny.png"/> How would you do it?</h2>
# MAGIC 
# MAGIC   <p>A significant number of problems stem from problems "on disk".</p>
# MAGIC 
# MAGIC   <p>When we are looking at the files "on disk" every system requires different techniques.</p>
# MAGIC 
# MAGIC   <p><b>Typically we need to answer two questions:</b></p>
# MAGIC   <li>How big are the files?</li>
# MAGIC   <li>How many files are there?</li>
# MAGIC 
# MAGIC   <p><b>Different Systems:</b></p>
# MAGIC   <li>Object Stores</li>
# MAGIC   <li>NAS</li>
# MAGIC   <li>HDFS</li>
# MAGIC   <li>Locally mounted file systems</li>
# MAGIC 
# MAGIC   <p><b>Different Techniques:</b></p>
# MAGIC   <li>Cloud Tools</li>
# MAGIC   <li>OS Tools</li>
# MAGIC   <li>Language Tools: Scala, Python, etc</li>
# MAGIC     
# MAGIC </div>  

# COMMAND ----------

# MAGIC %md-sandbox
# MAGIC ##![Spark Logo Tiny](https://files.training.databricks.com/images/105/logo_spark_tiny.png) The Introduction
# MAGIC 
# MAGIC When we do this on Databricks, we have our own set of tricks.
# MAGIC * How you do this will vary from system to system
# MAGIC * But the "who", "what", "where", <span style="text-decoration:line-through">how</span>, and "why" applies to everyone
# MAGIC * Today we are on Databricks, so we need to at least cover the Databricks techniques for the sake of this class
# MAGIC 
# MAGIC Let's start by exploring your cluster's file system.
# MAGIC 
# MAGIC A quick and easy way to do this is to shell out to the Driver and issue simple commands like **ls -la *target***

# COMMAND ----------

# MAGIC %sh ls -la /

# COMMAND ----------

# MAGIC %md
# MAGIC 
# MAGIC Visible on the Driver's root is the Databricks File System (DBFS).
# MAGIC 
# MAGIC The DBFS is a virtual file system hosted in an Azure Blob Store.
# MAGIC 
# MAGIC You can see it in the output above as **dbfs**, about the 7th item from the top.

# COMMAND ----------

# MAGIC %md
# MAGIC 
# MAGIC You can read more about the DBFS here:
# MAGIC * <a href="https://docs.azuredatabricks.net/user-guide/dbfs-databricks-file-system.html#dbfs" target="_blank">https&#58;//docs.azuredatabricks.net/user-guide/dbfs-databricks-file-system.html#dbfs</a>

# COMMAND ----------

# MAGIC %md
# MAGIC Of particular interest is the other object stores mounted to DBFS.
# MAGIC 
# MAGIC We can dig further into the contents of the DBFS, and specifically the **training** mount, with the following command:

# COMMAND ----------

# MAGIC %sh ls -la /dbfs/mnt/training

# COMMAND ----------

# MAGIC %md
# MAGIC The Databricks notebooks provide several other techniques for working with the file system.
# MAGIC 
# MAGIC The next is **%fs ls *target***

# COMMAND ----------

# MAGIC %fs ls dbfs:/mnt/training

# COMMAND ----------

# MAGIC %md
# MAGIC This is actually shorthand for the following piece of code:

# COMMAND ----------

path = 'dbfs:/mnt/training'
files = dbutils.fs.ls(path)

display(files)

# COMMAND ----------

# MAGIC %md
# MAGIC 
# MAGIC The **dbutils.fs.ls(..)** command returns a sequence in Scala (or a list in Python) containing **FileInfo** objects.
# MAGIC 
# MAGIC Of special interest to us are its attributes:
# MAGIC 
# MAGIC | Attribute   | Type      | Description |
# MAGIC |-------------|-----------|-------------|
# MAGIC | **path**    | string    | The path of the file or directory. |
# MAGIC | **name**    | string    | The name of the file or directory. |
# MAGIC | **isDir()** | boolean   | True if the path is a directory. |
# MAGIC | **size**    | long/int64| The length of the file in bytes or zero if the path is a directory. |

# COMMAND ----------

# MAGIC %md
# MAGIC 
# MAGIC For more information on **FileInfo** objects, you can see the Databricks documentation here:  
# MAGIC * <a href="https://docs.azuredatabricks.net/api/latest/dbfs.html#dbfsfileinfo" target="_blank">https&#58;//docs.azuredatabricks.net/api/latest/dbfs.html#dbfsfileinfo</a>

# COMMAND ----------

print('Collection Type: {}'.format(type(files)))

first = files[0]
print('Contained Type:  {}'.format(type(first)))

# COMMAND ----------

# MAGIC %md
# MAGIC We can use this information to programmatically explore the file system.

# COMMAND ----------

# MAGIC %md
# MAGIC As a side note, there are actually a number of different commands for working with the file system.
# MAGIC 
# MAGIC See **dbutils.fs.help** as show below:  

# COMMAND ----------

dbutils.fs.help()

# COMMAND ----------

# MAGIC %md
# MAGIC ##![Spark Logo Tiny](https://files.training.databricks.com/images/105/logo_spark_tiny.png) The Challenge
# MAGIC 
# MAGIC **You have 10 minutes to determine how many individual files make up our parquet file and its total size:**
# MAGIC * `dbfs:/mnt/training/asa/flights/1990-1999-by-year.parquet`
# MAGIC 
# MAGIC ** Requirements:**
# MAGIC * Report the two pieces of information as 
# MAGIC   * **totalCount**: The total number of files (but not directories) 
# MAGIC   * **totalSize**: The sum of the size of all files 
# MAGIC * The operation/technique should recurse into all subdirectories
# MAGIC 
# MAGIC ###Strategies:
# MAGIC * **Option #1**: Use **%fs ls dbfs:/mnt/training/asa/flights/1990-1999-by-year.parquet**, manually count every file, repeat for every subdirectory, and then add them all up.
# MAGIC * **Option #2**: Use **dbutls.fs.ls(..)** to programmatically sum and/or count the number of files in each directory.
# MAGIC   * **2.A:** Use Scala's native functions (or Pythons) to sum the count of all files.
# MAGIC   * **2.B:** Convert the raw data to a DataFrame and use the DataFrame's API to sum the count of all files.
# MAGIC * **Option #3** Use **%sh** to run linux commands like **du** (which sadly doesn't work)
# MAGIC * **Option #4**: Surprise me!
# MAGIC 
# MAGIC ###Bonus
# MAGIC * Wrap it all up into a re-usable function.

# COMMAND ----------

# ANSWER

def computeFileStats(path):
  bytes = 0
  count = 0

  files = dbutils.fs.ls(path)
 
  while (len(files) > 0):
    fileInfo = files.pop(0)
    if (fileInfo.isDir() == False):
      count += 1
      bytes += fileInfo.size
    else:
      files.extend(dbutils.fs.ls(fileInfo.path))  # append multiple object to files
      
  return (count, bytes)

# COMMAND ----------

# ANSWER
path = 'dbfs:/mnt/training/asa/flights/1990-1999-by-year.parquet'

(totalCount, totalBytes) = computeFileStats(path)

print("{:,} files".format(totalCount))
print("{:,} bytes".format(totalBytes))

# COMMAND ----------

# MAGIC %md
# MAGIC ##![Spark Logo Tiny](https://s3-us-west-2.amazonaws.com/curriculum-release/images/105/logo_spark_tiny.png) Validate Your Answer
# MAGIC Run the cell below to validate your answer:

# COMMAND ----------

clearYourResults()

validateYourAnswer("01) Total Count", 1680788708, totalCount)
validateYourAnswer("02) Total Bytes", 1928383137, totalBytes)

summarizeYourResults()

# COMMAND ----------

# MAGIC %md
# MAGIC ##![Spark Logo Tiny](https://s3-us-west-2.amazonaws.com/curriculum-release/images/105/logo_spark_tiny.png) Our Utility Methods
# MAGIC 
# MAGIC Before moving on, let's take a look at some of the utility methods we created just for this class:

# COMMAND ----------

# MAGIC %run "../Includes/Utility Methods"

# COMMAND ----------

# MAGIC %md
# MAGIC ### Compute File Stats
# MAGIC 
# MAGIC Given a path, returns the number of files and the total size on disk.
# MAGIC 
# MAGIC Useful for diagnosing a number of on-disk problems

# COMMAND ----------

path = "dbfs:/mnt/training/asa/flights/1990-1999-by-year.parquet"

(count, bytes) = computeFileStats(path)

print("{:14,} files".format(count))
print("{:14,} bytes".format(bytes))
print("{:14,} KB/file".format(int(bytes/count/1024)))

print("-"*80)

# COMMAND ----------

# MAGIC %md
# MAGIC ### Print Records Per Partition
# MAGIC Given a DataFrame, this method will count the number of records in each Spark-Partition.
# MAGIC 
# MAGIC Usefull for identifying things like in-memory skew

# COMMAND ----------

path = "dbfs:/mnt/training/asa/flights/1990-1999-by-year.parquet"
df = spark.read.parquet(path)

printRecordsPerPartition(df)

# COMMAND ----------

# MAGIC %md
# MAGIC ### Benchmark Count
# MAGIC 
# MAGIC A quick & dirty benchmarking tool
# MAGIC 
# MAGIC Used to evaluate the performance difference between different iterations of code.

# COMMAND ----------

path = "dbfs:/mnt/training/asa/flights/1990-1999-by-year.parquet"
dfA = spark.read.parquet(path)

(dfB, total, duration) = benchmarkCount(lambda: df)

print("Query Duration: {:,} ms".format(int(duration)))
print("-"*80)

# COMMAND ----------

# MAGIC %md-sandbox
# MAGIC &copy; 2018 Databricks, Inc. All rights reserved.<br/>
# MAGIC Apache, Apache Spark, Spark and the Spark logo are trademarks of the <a href="http://www.apache.org/">Apache Software Foundation</a>.<br/>
# MAGIC <br/>
# MAGIC <a href="https://databricks.com/privacy-policy">Privacy Policy</a> | <a href="https://databricks.com/terms-of-use">Terms of Use</a> | <a href="http://help.databricks.com/">Support</a>
